Renaming to match Arduino convention
